[Chore #165051474] Move signup and signin tests into authorization.test.js file
