+++ server/test/account.test.js
@@ -1,41 +1,64 @@
 // import chai from 'chai'
 // import chaiHttp from 'chai-http'
 // import app from '../app'
-// import request from 'supertest'
 
 // const { expect } = chai
 // chai.use(chaiHttp)
 
-// let token
-//   //now let's login the user before we run any tests
-//   before("User Signup", (done) => {
-//     const user = {
-//         firstName : "Joseph",
-//         lastName : "joe",
-//         email: 'joe@gmail.com', 
-//         password: '123456'
-//       }
-//       chai.request(app)
-//       .post('/api/v1/auth/signup')
-//       .set("Content-type","application/json")
-//       .set("Accept","application/json")
-//       .send(user)
-//       .end((err, res) =>{
-//           console.log(res.body)
-//         token = res.body.token
-//         expect(res.body.status).to.equal(201)
-//         expect(res.body).to.have.property('status')
-//         expect(res.body).to.have.property('message')
-//         expect(res.body).to.have.property('data')
-//         expect(res.body).to.be.an('object')
-//         done();
-//       });
-//   });
+// chai.use(chaiHttp);
+// chai.use(chaiHttp);
+
+// let token;
+
+// before("User Login",(done)=>{
+//         const user = {
+//           email: "joe@gmail.com",
+//           password: "123456"
+//         };
+//     chai.request(app)
+//     .post("/api/v1/auth/signin")
+//     .set("Content-type","application/json")
+//     .set("Accept","application/json")
+//     .send(user)
+//     .end((err,res)=>{
+//         if(err){
+//             done(err)
+//         }
+//         token=res.body.token;
+//         expect(res.body).to.be.an("object")
+//         expect(res.body).to.have.property("status")
+//         expect(res.body).to.have.property("error")
+//         expect(res.body.status).to.be.equal(400)
+//         done()
+//     })
+// })
 
 // describe('Create account', () =>{
-//     it('An authorized', () =>{
+//   it('Correct credentials', (done) =>{
+//     chai.request(app)
+//     .post("/api/v1/auth/signin")
+//     .send({
+//         email: "joe@gmail.com",
+//         password:"123456"
+//     })
+//     .end((err,res)=>{
+//         if(err){
+//             done(err);
+//         }
+//         expect(res.body).to.be.an("object");
+//         expect(res.body).to.have.property("status");
+//         expect(res.body).to.have.property("message");
+//         expect(res.body).to.have.property("data");
+//         expect(res.body.status).to.be.equal(200);
+//         done();
+//     })
+//   })
+
+//     it('Create account', () =>{
 //         chai.request(app)
 //         .post('/api/v1/accounts')
+//         .set("Content-type","application/json")
+//         .set("Authorization",token)
 //         .send({
 //             accountNumber : 123456,
 //             firstName : "Joseph",
